Feat/gunnraj/#94 saved chat history (#97)
# Task
Resolves #94  - Implement Chat History

# Root Cause
Chat history is not saved permanantly, it is only in localStorage in the
browser.

# Solution 
The flow is now as follows:
When a new project is made:
- chat history is checked, if empty, the initlal ai message is shown on
the ui and saved to the backend
- when a user sends a message, this message is shown on the ui and saved
to the backend
- when the ai generates a response, this is saved to the backend then
sent to the front end

We check the current project vs the AI reponse requested project id, if
they dont match, then it is not shown on the front end. This prevents
the issue where changing projects before the ai response is generated
causes the chat to shown the previous project's messages.

IMPORTANT: The online ai endpoint also seemed to be deprecated, so i
changed it to a new one.

# How Has This Been Tested? 
Ad Hoc testing.

# Checklist: 
- [✅] I have linked issue to this Pull Request 
- [✅] I have performed a self-review of my code 
- [✅] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation 
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes

---------

Co-authored-by: Gunnraj98 <gdha0007@student.monash.edu>