Issue #90 : Added Currently Working title and description of project visibility in NavBar (#104)
…y of current projcet in use. GitHub Issue #90.

# Task
Resolves #ISSUE 90

# Root Cause
<!--- Please describe why this change is required. What's the motive?
-->
This change was required as it was not easy to identify which project
was being worked on. As well as no where to see description of project
after you create a project.

# Solution 
<!--- Please describe your code changes in detail for developers. --> 
edited NavBar.jsx to add a new little shell next to Qaul AI logo. 

# How Has This Been Tested? 
<!--- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce. Please also list any relevant
details for your test configuration -->
I ran the program to see if the title was changing according to project
and added descriptions to projects to see if the description of the
project is also visible.

# Checklist: 
- [ ] I have linked issue to this Pull Request 
- [ ] I have performed a self-review of my code 
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation 
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes

Co-authored-by: chap989 <rcha0087@student.monash.edu>